requirements.txt
cityscapesscripts==2.2.4
clip==1.0
dassl==0.6.3
ftfy==6.3.1
fvcore==0.1.5.post20221221
horovod==0.28.1
huggingface_hub==0.21.4
ipython==8.12.3
loralib==0.1.2
matplotlib==3.7.3
mmcv==1.7.0
mmcv_full==1.7.1
mmdet==3.3.0
mmengine==0.10.1
mpi4py==3.1.4
numpy==1.24.3
opencv_python==4.8.0.76
opencv_python_headless==4.8.1.78
panopticapi==0.1
Pillow==10.0.0
Pillow==11.3.0
pycocotools==2.0.7
pysodmetrics==1.4.2
PyYAML==6.0.1
PyYAML==6.0.2
regex==2023.12.25
safetensors==0.5.3
scikit_learn==1.3.0
scipy==1.16.0
setuptools==59.5.0
tensorboardX==2.6.2
tensorboardX==2.6.4
terminaltables==3.1.10
timm==1.0.16
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.31.0
